Move model quantization helpers to shared utility
