+++ src/megatron/bridge/models/kimi_vl/utils.py
@@ -12,128 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-
-import torch
-
-
-def dequantize_int4(
-    weight_packed: torch.Tensor,
-    weight_scale: torch.Tensor,
-    weight_shape: torch.Tensor,
-    group_size: int = 32,
-    device: str | torch.device | None = None,
-) -> torch.Tensor:
-    """Dequantize INT4 packed weights to bfloat16.
-
-    Extracts local tensors from DTensors before unpacking (bitwise ops don't work on DTensor).
-    Both weight_packed and weight_scale should have matching sharding so .to_local() gives
-    corresponding slices automatically.
-
-    Args:
-        weight_packed: INT4 packed weights [out_features, in_features // 8], may be DTensor
-        weight_scale: Per-group scales [out_features, num_groups], should be DTensor with same sharding
-        weight_shape: Original shape [2], stores global dimensions
-        group_size: Elements per scale group (default 32)
-        device: Target device for computation
-    """
-
-    local_out, local_packed_in = weight_packed.shape
-    local_in = local_packed_in * 8  # 8 INT4 values per int32
-
-    target_device = weight_packed.device if device is None else torch.device(device)
-    use_cuda = target_device.type == "cuda" and torch.cuda.is_available()
-
-    if use_cuda:
-        weight_packed = weight_packed.to(target_device)
-        weight_scale = weight_scale.to(target_device)
-
-    # Unpack INT4: [out, packed_in] -> [out, packed_in, 8] -> [out, in_features]
-    shifts = torch.arange(8, device=weight_packed.device) * 4
-
-    packed_unsqueezed = weight_packed.unsqueeze(-1)
-    unpacked = ((packed_unsqueezed >> shifts) & 0xF).float()
-    unpacked = unpacked.reshape(local_out, local_in)
-
-    # Convert unsigned 4-bit (0-15) to signed (-8 to 7) using OFFSET BINARY
-    # This matches compressed-tensors library which packs as: value + 8
-    # So unpack as: value - 8
-    unpacked = unpacked - 8
-
-    # Apply scale - both are now local tensors with corresponding slices
-    scale = weight_scale.float()
-    if scale.ndim == 1:
-        local_num_groups = scale.numel() // local_out
-        scale = scale.view(local_out, local_num_groups)
-    else:
-        scale = scale.view(local_out, -1)
-
-    local_num_groups = scale.shape[1]
-    elements_per_group = local_in // local_num_groups
-
-    # repeat_interleave expands [local_out, local_num_groups] -> [local_out, local_in]
-    scale_expanded = scale.repeat_interleave(elements_per_group, dim=1)
-
-    if scale_expanded.shape[1] < local_in:
-        # Pad if needed
-        scale_expanded = torch.nn.functional.pad(
-            scale_expanded, (0, local_in - scale_expanded.shape[1]), value=scale_expanded[:, -1:].mean()
-        )
-    scale_expanded = scale_expanded[:, :local_in]
-    result = unpacked * scale_expanded
-
-    result = result.to(torch.bfloat16)
-
-    return result
+from megatron.bridge.models.conversion.quantization_utils import dequantize_int4, quantize_to_int4
 
 
-def quantize_to_int4(
-    weight: torch.Tensor,
-    group_size: int = 32,
-    scale_dtype: torch.dtype = torch.bfloat16,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    """Quantize bfloat16/float16 weights to INT4 packed format.
-
-    Returns:
-        weight_packed: INT4 values packed into int32 (8 values per int32)
-        weight_scale: Per-group scale factors in ``scale_dtype``
-        weight_shape: Original tensor shape (int32)
-    """
-    out_features, in_features = weight.shape
-    weight_shape = torch.tensor([out_features, in_features], dtype=torch.int32)
-
-    # Convert to float32 for computation
-    w = weight.float()
-
-    # Compute per-group scales (group along in_features dimension)
-    num_groups = (in_features + group_size - 1) // group_size
-    w_grouped = w.view(out_features, num_groups, -1)
-
-    # Symmetric quantization: scale = max(abs(w)) / 7
-    group_max = w_grouped.abs().amax(dim=-1)
-    scale = group_max / 7.0
-    scale = scale.clamp(min=1e-10)  # Avoid division by zero
-
-    # Quantize: w_q = round(w / scale), clamp to [-8, 7]
-    scale_expanded = scale.unsqueeze(-1).expand_as(w_grouped)
-    w_q = (w_grouped / scale_expanded).round().clamp(-8, 7)
-
-    # Convert signed [-8, 7] to unsigned [0, 15] using offset binary
-    w_q = w_q.view(out_features, -1)[:, :in_features]
-    w_q = (w_q + 8).to(torch.uint8)
-
-    # Pack 8 INT4 values into each int32 along the in_features dimension
-    # HF format: [out_features, in_features//8] - 2D packed tensor
-    assert in_features % 8 == 0, f"in_features must be divisible by 8, got {in_features}"
-
-    # Reshape to [out_features, in_features//8, 8] for packing along dim 1
-    w_q_grouped = w_q.view(out_features, in_features // 8, 8).to(torch.int32)
-
-    # Pack 8 nibbles into 1 int32
-    packed = torch.zeros(out_features, in_features // 8, dtype=torch.int32, device=weight.device)
-    for i in range(8):
-        packed |= (w_q_grouped[:, :, i] & 0xF) << (i * 4)
-
-    weight_packed = packed
-    weight_scale = scale.to(scale_dtype)
-
-    return weight_packed, weight_scale, weight_shape
+__all__ = ["dequantize_int4", "quantize_to_int4"]
